// store/posts/types.ts
export interface Post {
  id: number;
  title: string;
  body: string;
  userId: number;
}

export interface PostsState {
  list: Post[];
  currentPost: Post | null;
  loading: boolean;
  error: string | null;
}

export enum PostsActionTypes {
  FETCH_POSTS_REQUEST = "posts/FETCH_POSTS_REQUEST",
  FETCH_POSTS_SUCCESS = "posts/FETCH_POSTS_SUCCESS",
  FETCH_POSTS_FAILURE = "posts/FETCH_POSTS_FAILURE",

  FETCH_POST_BY_ID_REQUEST = "posts/FETCH_POST_BY_ID_REQUEST",
  FETCH_POST_BY_ID_SUCCESS = "posts/FETCH_POST_BY_ID_SUCCESS",
  FETCH_POST_BY_ID_FAILURE = "posts/FETCH_POST_BY_ID_FAILURE",

  CREATE_POST_REQUEST = "posts/CREATE_POST_REQUEST",
  CREATE_POST_SUCCESS = "posts/CREATE_POST_SUCCESS",
  CREATE_POST_FAILURE = "posts/CREATE_POST_FAILURE",

  UPDATE_POST_REQUEST = "posts/UPDATE_POST_REQUEST",
  UPDATE_POST_SUCCESS = "posts/UPDATE_POST_SUCCESS",
  UPDATE_POST_FAILURE = "posts/UPDATE_POST_FAILURE",

  DELETE_POST_REQUEST = "posts/DELETE_POST_REQUEST",
  DELETE_POST_SUCCESS = "posts/DELETE_POST_SUCCESS",
  DELETE_POST_FAILURE = "posts/DELETE_POST_FAILURE",
}